[HBO-5] Rename PokemonFragment2.kt on NewPokemonFragment refactor init lazy adapter refactor MutableListof on ArrayList
